+++ apm-agent-plugins/apm-jmx-plugin/src/main/java/co/elastic/apm/agent/jmx/JmxConfiguration.java
@@ -34,7 +34,7 @@ public class JmxConfiguration extends ConfigurationOptionProvider {
 
     private ConfigurationOption<List<JmxMetric>> captureJmxMetrics = ConfigurationOption.<List<JmxMetric>>builder(JmxMetric.TokenValueConverter.INSTANCE, List.class)
         .key("capture_jmx_metrics")
-        .tags("added[1.9.0]")
+        .tags("added[1.10.0]")
         .description("Report metrics from JMX to the APM Server\n" +
             "\n" +
             "Can contain multiple comma separated JMX metric definitions:\n" +